Superstore Financials Analysis
This report provides an in-depth analysis of the financial performance of a fictional store, called Superstore. It includes key metrics like Total Revenue, Profit, Units Sold and Discounts across 5 different countries (Mexico, France, USA, Germany and Canada) and years (2014 and 2013). The main objective of the report is to show the KPIs and their performance, mainly during the year 2014, and using its 2013 KPIs to show a comparison in performance.
The analysis is visualized through the use of pie charts, bar charts, KPI cards and Tool Tips that improve the visualization and understanding of the results.
The report is designed for sales managers, business analysts, and executives to monitor financial performance and make data-driven decisions.